+++ packages/nimble-components/src/breadcrumb-item/tests/breadcrumb-item.react.tsx
@@ -1,4 +1,0 @@
-import { BreadcrumbItem } from '..';
-import { wrap } from '../../utilities/tests/react-wrapper';
-
-export const NimbleBreadcrumbItem = wrap(BreadcrumbItem);