+++ packages/nimble-components/src/table-column/text/tests/table-column-text.react.tsx
@@ -1,4 +1,0 @@
-import { TableColumnText } from '..';
-import { wrap } from '../../../utilities/tests/react-wrapper';
-
-export const NimbleTableColumnText = wrap(TableColumnText);