+++ packages/nimble-components/src/table/tests/table.react.tsx
@@ -1,4 +1,0 @@
-import { Table } from '..';
-import { wrap } from '../../utilities/tests/react-wrapper';
-
-export const NimbleTable = wrap(Table);