+++ packages/nimble-components/src/toolbar/tests/toolbar.react.tsx
@@ -1,4 +1,0 @@
-import { Toolbar } from '..';
-import { wrap } from '../../utilities/tests/react-wrapper';
-
-export const NimbleToolbar = wrap(Toolbar);