+++ packages/storybook/src/nimble/rich-text/viewer/rich-text-viewer.react.tsx
@@ -1,0 +1,4 @@
+import { RichTextViewer } from '@ni/nimble-components/dist/esm/rich-text/viewer';
+import { wrap } from '../../../utilities/react-wrapper';
+
+export const NimbleRichTextViewer = wrap(RichTextViewer);