+++ packages/storybook/src/nimble/tabs-toolbar/tabs-toolbar.react.tsx
@@ -1,0 +1,4 @@
+import { TabsToolbar } from '@ni/nimble-components/dist/esm/tabs-toolbar';
+import { wrap } from '../../utilities/react-wrapper';
+
+export const NimbleTabsToolbar = wrap(TabsToolbar);