+++ packages/storybook/src/nimble/text-field/text-field.react.tsx
@@ -1,0 +1,4 @@
+import { TextField } from '@ni/nimble-components/dist/esm/text-field';
+import { wrap } from '../../utilities/react-wrapper';
+
+export const NimbleTextField = wrap(TextField);