+++ packages/storybook/src/nimble/tree-view/tree-view.react.tsx
@@ -1,0 +1,4 @@
+import { TreeView } from '@ni/nimble-components/dist/esm/tree-view';
+import { wrap } from '../../utilities/react-wrapper';
+
+export const NimbleTreeView = wrap(TreeView);